Changed project structure and fixed the flow.
