+++ pom.xml
@@ -33,8 +33,7 @@
         <dependency>
             <groupId>org.testng</groupId>
             <artifactId>testng</artifactId>
-            <version>6.9.4</version>
-            <scope>test</scope>
+            <version>7.0.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
