+++ src/main/resources/configuration.properties
@@ -1,0 +1,1 @@
+ELASTIC_SEARCH_URL=http://localhost:9200/app/suite
